adblib: Add support for JDWP process properties
Add support for opening JDWP connection to device processes,
and fetch some useful DDMS packets from each process, such
as "HELO", "APNM", "REAQ", "WAIT".

Bug: 188067646, 237420199
Test: Included

+++ adblib-tools/src/com/android/adblib/tools/debugging/packets/JdwpPacketUtils.kt
@@ -163,7 +163,7 @@ internal fun MutableJdwpPacket.parseHeader(buffer: ByteBuffer) {
     // Byte [10, 10] : cmd     || (if reply) error code low byte
     packetLength = buffer.getInt()
     packetId = buffer.getInt()
-    packetFlags = buffer.get().toInt()
+    packetFlags = buffer.get().toUByte().toInt()
     if (isReply) {
         errorCode = buffer.getShort().toUShort().toInt()
     } else {
